fixes, organisation
