+++ COTL_API/Helpers/BaseClassReversePatches.cs
@@ -1,0 +1,16 @@
+﻿using HarmonyLib;
+using System;
+using System.Runtime.CompilerServices;
+
+namespace COTL_API.Helpers;
+
+public static class OnInteractHelper
+{
+    [HarmonyPatch(typeof(Interaction), nameof(Interaction.OnInteract))]
+    [HarmonyReversePatch]
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public static void Interaction_OnInteract(Interaction instance, StateMachine state)
+    {
+       if(Plugin.Debug) Plugin.Logger.LogWarning($"Interaction.OnInteract Test({instance}, {state})");
+    }
+}
